requirements.txt
playwright
faiss-cpu
sentence-transformers
slack-bolt
